feat: cli to export experiment SB3s
The SB3 zip for all users of an experiment may get very big. Therefore,
the export via the browser takes too long. Instead, this feature allows
starting the application with additional CLI flags  as shown below that
trigger a direct export to a file on the server.

Start ScratchLog like
```
java -Dserver.port=0 scratchLog.jar sb3export ./outputDir/ 1 2 3
```
with `1`, `2`, `3` being the IDs of the experiments for which all SB3
files should be exported.
`-Dserver.port=0` uses a random unused port to start the application, so
that it can be started even with an actual ScratchLog instance already
running on the same server on the regular port.